requirements.txt
asttokens==2.4.1
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.6.2
charset-normalizer==2.0.4
comm==0.2.2
contourpy==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.51.0
gmpy2==2.1.2
idna==3.7
importlib_metadata==8.0.0
importlib-resources==6.1.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.2.1
numexpr==2.8.7
numpy==1.26.4
onnx==1.16.1
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
ply==3.11
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.32.2
scipy==1.13.1
setuptools==69.5.1
sip==6.7.12
six==1.16.0
stack-data==0.6.2
sympy==1.12
tomli==2.0.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.2.0
typing_extensions==4.11.0
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.43.0
zipp==3.17.0
